backend/requirements.txt
# ──────────────────────────────────────────────────────────
# Menstrual Health AI/ML Pipeline — Python Dependencies
# ──────────────────────────────────────────────────────────
# Install with:
#   pip install -r requirements.txt

# Core ML & data science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0

# API server (for Flutter integration)
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.9

# Environment / config
python-dotenv>=1.0.0

# Security / Auth
bcrypt>=4.0.0              # password hashing
python-jose[cryptography]>=3.3.0  # JWT tokens

# MongoDB Atlas
pymongo>=4.7.0
dnspython>=2.6.0        # required for mongodb+srv:// URIs

# Firebase (push notifications — service account in firebase-adminsdk.json)
firebase-admin>=7.0.0

# HTTP client (for product APIs)
httpx>=0.27.0

# Optional: data exploration / debugging
# jupyter
# matplotlib
# seaborn
